feat: inspector overhaul + detail panel + canvas smoothing
Inspector:
- Single-line toolbar: search (with #id exact match), type/sort
  combos, group toggle, right-aligned filtered/total count
- Replace flat TreeNode list with ImGui Table (Type/Name/Pos,
  tinted Type cell using entity color, span-all-columns
  Selectable)
- Sort modes: default / ID / Type / name / distance-to-player
- Group-by-type mode (CollapsingHeader per type)
- Section the detail pane via CollapsingHeader (basic info,
  same-cell entities, selected cell, entities on cell)
- Cell list now has search box + group-by-type
- Make x/y/radius read-only (debug UI is for viewing)

Canvas:
- Float-precision camera tracking (cam_x/y as float, follow
  uses entity fx/fy) — fixes sub-cell jitter
- Render entities at fx/fy so the selection outline glides
  with the entity instead of snapping
- View boundary outline tracks selected entity smoothly
- Selected-cell outline now cyan, distinguishable from
  entity yellow

Detail panel:
- New DrawEntityDetail(world) panel
- New RegisterEntityDetailText(type, fn) extension point —
  callback returns pre-formatted multi-line text, rendered
  in a scrollable child

Demo / docs:
- Default-select player on world init (selected_id = player_id)
- main.cpp: vx/vy read-only display + Player/停止移动 command
  example demonstrating "modify via command, not via Inspector"
- README: new section + directory listing updates

+++ src/dui_canvas.h
@@ -10,8 +10,8 @@ struct CanvasView {
     bool show_labels   = true;
     bool show_axis     = true;
     bool follow_player = true;
-    int  cam_x = 0;
-    int  cam_y = 0;
+    float cam_x = 0.f;
+    float cam_y = 0.f;
 };
 
 // view == nullptr uses an internal static default instance.
